Used rtree to get the nearest geographic object
